Fixed: Ulong typedef truncation on Windows 64-bit in formatdata.c
Changed typedef from unsigned long to uintptr_t. Ulong is used for
pointer casts throughout the file, and on Windows LLP64 unsigned long
is 32-bit while pointers are 64-bit.

+++ src/formatdata.c
@@ -31,7 +31,7 @@ use or performance of this software.
 
 static int memno2info Argdcl((int, Namep*));
 
-typedef unsigned long Ulong;
+typedef uintptr_t Ulong;
 
  extern char *initbname;
